requirements.txt
Flask
tensorflow
numpy
Pillow
pandas
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Cors
PyMySQL
python-dotenv 
Werkzeug